added various stuff

+++ SafeBoxes_TS2_ALAFAK/Frm_Clients.vb
@@ -19,13 +19,14 @@ Public Class Frm_Clients
 
     Private Sub btn_delclient_Click(sender As Object, e As EventArgs) Handles btn_delclient.Click
         InputBox.Show("Enter admin password:", "Delete Client#" & dgv_clients.SelectedRows(0).Cells(0).Value, True)
-        If inResult = "12345" Then
-            ExecuteQuery("DELETE FROM Clients WHERE ClientId = " & dgv_clients.SelectedRows(0).Cells(0).Value)
-            FillDGV(dgv_clients, "SELECT ClientId AS [ID], ClientFName AS [First Name], ClientMName AS [Father Name], ClientLName AS [Last Name], ClientDOB AS [Date Of Birth], ClientMother AS [Mother Name], ClientRegisterNbr, PostBoxNbr,RegionName  FROM Clients c,Regions r WHERE c.RegionId=r.RegionId")
-        Else
-            MessageBox.Show("Wrong password!")
+        If Not inResult.Equals("0") Then
+            If inResult = "12345" Then
+                ExecuteQuery("DELETE FROM Clients WHERE ClientId = " & dgv_clients.SelectedRows(0).Cells(0).Value)
+                FillDGV(dgv_clients, "SELECT ClientId AS [ID], ClientFName AS [First Name], ClientMName AS [Father Name], ClientLName AS [Last Name], ClientDOB AS [Date Of Birth], ClientMother AS [Mother Name], ClientRegisterNbr, PostBoxNbr,RegionName  FROM Clients c,Regions r WHERE c.RegionId=r.RegionId")
+            Else
+                MessageBox.Show("Wrong password!")
+            End If
         End If
-
     End Sub
 
     Private Sub dgv_clients_SelectionChanged(sender As Object, e As EventArgs) Handles dgv_clients.SelectionChanged
