+++ SafeBoxes_TS2_ALAFAK/Frm_Contracts.vb
@@ -28,22 +28,23 @@
 
     Private Sub btn_delcontract_Click(sender As Object, e As EventArgs) Handles btn_delcontract.Click
         InputBox.Show("Enter admin password:", "Delete Contract#" & dgv_contracts.SelectedRows(0).Cells(0).Value, True)
-        If inResult = "12345" Then
-            ExecuteQuery("DELETE FROM Contract WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM SignatureCards WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM Permissions WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM ContEnd WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM Visits WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM MissingKeys WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM InfoVoucher WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM Authorizations WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM Renew WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            ExecuteQuery("DELETE FROM FATCA WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
-            FillDGV(dgv_contracts, "SELECT ContId AS ID, ContBDate AS [Contract Date], ContToDate AS [Expire Date], ContPhone1 AS [Phone1], ContPhone2 AS [Phone2], BoxId AS [Box ID], AccountId AS [Account ID], BuildingName AS Address, EmpFName + ' ' + EmpLName AS Employee FROM Contract, Buildings, Employees WHERE Contract.BuildingId = Buildings.BuildingId AND Contract.EmpId = Employees.EmpId AND ContId NOT IN (SELECT ContId FROM ContEnd) And ContToDate > date()")
-        Else
-            MessageBox.Show("Wrong password!")
+        If Not inResult.Equals("0") Then
+            If inResult = "12345" Then
+                ExecuteQuery("DELETE FROM Contract WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM SignatureCards WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM Permissions WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM ContEnd WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM Visits WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM MissingKeys WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM InfoVoucher WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM Authorizations WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM Renew WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                ExecuteQuery("DELETE FROM FATCA WHERE ContId = " & dgv_contracts.SelectedRows(0).Cells(0).Value)
+                FillDGV(dgv_contracts, "SELECT ContId AS ID, ContBDate AS [Contract Date], ContToDate AS [Expire Date], ContPhone1 AS [Phone1], ContPhone2 AS [Phone2], BoxId AS [Box ID], AccountId AS [Account ID], BuildingName AS Address, EmpFName + ' ' + EmpLName AS Employee FROM Contract, Buildings, Employees WHERE Contract.BuildingId = Buildings.BuildingId AND Contract.EmpId = Employees.EmpId AND ContId NOT IN (SELECT ContId FROM ContEnd) And ContToDate > date()")
+            Else
+                MessageBox.Show("Wrong password!")
+            End If
         End If
-
     End Sub
 
     Private Sub dgv_contracts_SelectionChanged(sender As Object, e As EventArgs) Handles dgv_contracts.SelectionChanged
